+++ music-api/src/main/java/music/mapper/TrackMapper.java
@@ -33,4 +33,6 @@ public interface TrackMapper {
     void markDeletedById(@Param("id")long id, @Param("deletedInd") boolean deletedInd);
 
     void setRatingById(long id, byte rating);
+
+    void updateFieldById(long id, String field, Object newValue, String type);
 }