+++ music-api/src/main/java/music/mapper/UpdateMapper.kt
@@ -9,4 +9,6 @@ interface UpdateMapper {
     fun insertUpdate(songId: Long, field: String, newValue: String)
     fun listByTrackId(songId: Long):List<TrackUpdate>
     fun list():List<TrackUpdate>
+	fun count():Long
+	fun deleteById(id:Long)
 }