+++ music-api/src/main/java/music/model/TrackUpdate.kt
@@ -1,9 +1,13 @@
 package music.model
 
 data class TrackUpdate(
-        val id: Long?,
+        val id: Long,
         val songId: Long,
         val field: String,
         val newValue: String,
         val updateType: Long?
-)
+) {
+	override fun toString(): String {
+		return "TrackUpdate(id=$id, songId=$songId, field='$field', newValue='$newValue', updateType=$updateType)"
+	}
+}