+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # changelog
 
+## 1.5-2019.11.19
+* feature: persist metadata changes for a track to disk
+
 ## 1.4-2019.11.18
 * feature: edit metadata for a track
+* fix: cannot change a track field if the current value is null
 
 ## 1.3-2019.11.15
 * feature: add endpoint for count of purgeable tracks (tracks marked deleted in database)